requirements.txt
streamlit==1.31.1
google-generativeai==0.7.2
langchain-google-genai==1.0.6
langchain==0.2.12
langchain-core==0.2.40
langchain-community==0.2.10
pillow==10.3.0
pytesseract==0.3.10
pyttsx3==2.90
gTTS==2.3.1
numpy==1.26.4
opencv-python-headless==4.8.1.82
python-dotenv==1.0.1
